Add files via upload
In this project I have used random module which will help us to generate random password. Firstly we will ask user to enter the length of password .And then we will use random.sample(s,k).Here s means the string which will contain (A-Z,a-z,0-9 and sample special symbol) and k means the length of password required. Using this function we will generate random password.